mainWindow: Make record button a "suggested-action"
Previously, the record button used the red "destructive" color for
the record button. As this is the main function of the application,
as well as being entirely nondestructive but in fact productive,
this wasn't sensible.

This has been changed from destructive-action to suggested-action,
as it is more in line with the button's role as the main purpose
of the application.

Closes #36

Signed-off-by: Felipe Borges <felipeborges@gnome.org>