Fix export dialog that disappears
When pressing the EXPORT button, the file dialog randomly
disappears before being able to specify a place where to store
the file. This is because the Gtk.FileChooser dialog is stored
in a local variable, which, as soon as the callback where it is
created returns, becomes unreferenced, and is reclaimed by the
garbage collector (which happens at random intervals). This
bug can be reproduced just by recording something, pressing
the "export" button, and waiting.

This patch fixes it by storing the dialog as a property in the
row object.

Fix https://gitlab.gnome.org/GNOME/gnome-sound-recorder/-/merge_requests/164